+++ src/gitManager/isomorphicGit.ts
@@ -112,7 +112,7 @@ export class IsomorphicGit extends GitManager {
                     // because that's what requestUrl expects
                     let collectedBody: ArrayBuffer | undefined;
                     if (body) {
-                        collectedBody = (await collect(body)).buffer;
+                        collectedBody = await asyncIteratorToArrayBuffer(body);
                     }
 
                     const res = await requestUrl({
@@ -122,11 +122,12 @@ export class IsomorphicGit extends GitManager {
                         body: collectedBody,
                         throw: false,
                     });
+
                     return {
                         url,
                         method,
                         headers: res.headers,
-                        body: [new Uint8Array(res.arrayBuffer)],
+                        body: arrayBufferToAsyncIterator(res.arrayBuffer),
                         statusCode: res.status,
                         statusMessage: res.status.toString(),
                     };
@@ -146,7 +147,7 @@ export class IsomorphicGit extends GitManager {
         }
     }
 
-    async status(): Promise<Status> {
+    async status(opts?: { path?: string }): Promise<Status> {
         let notice: Notice | undefined;
         const timeout = window.setTimeout(() => {
             notice = new Notice(
@@ -156,21 +157,34 @@ export class IsomorphicGit extends GitManager {
         }, 20000);
         try {
             this.plugin.setPluginState({ gitAction: CurrentGitAction.status });
+            const statusOpts = { ...this.getRepo() } as Parameters<
+                typeof git.statusMatrix
+            >[0];
+            if (opts?.path != undefined) {
+                statusOpts.filepaths = [`${opts.path}/`];
+            }
             const status = (
-                await this.wrapFS(git.statusMatrix({ ...this.getRepo() }))
+                await this.wrapFS(git.statusMatrix(statusOpts))
             ).map((row) => this.getFileStatusResult(row));
 
-            const changed = status.filter(
-                (fileStatus) => fileStatus.workingDir !== " "
-            );
-            const staged = status.filter(
-                (fileStatus) =>
-                    fileStatus.index !== " " && fileStatus.index !== "U"
-            );
+            const changed: FileStatusResult[] = [];
+            const staged: FileStatusResult[] = [];
+            const all: FileStatusResult[] = [];
+            for (const file of status) {
+                if (file.workingDir !== " ") {
+                    changed.push(file);
+                }
+                if (file.index !== " " && file.index !== "U") {
+                    staged.push(file);
+                }
+                if (file.index != " " || file.workingDir != " ") {
+                    all.push(file);
+                }
+            }
             const conflicted: string[] = [];
             window.clearTimeout(timeout);
             notice?.hide();
-            return { all: status, changed, staged, conflicted };
+            return { all, changed, staged, conflicted };
         } catch (error) {
             window.clearTimeout(timeout);
             notice?.hide();
@@ -287,8 +301,8 @@ export class IsomorphicGit extends GitManager {
                 const filesToStage =
                     unstagedFiles ?? (await this.getUnstagedFiles(dir ?? "."));
                 await Promise.all(
-                    filesToStage.map(({ path, deleted }) =>
-                        deleted
+                    filesToStage.map(({ path, type }) =>
+                        type == "D"
                             ? git.remove({ ...this.getRepo(), filepath: path })
                             : this.wrapFS(
                                   git.add({ ...this.getRepo(), filepath: path })
@@ -370,13 +384,20 @@ export class IsomorphicGit extends GitManager {
         if (status) {
             if (dir != undefined) {
                 files = status.changed
-                    .filter((file) => file.path.startsWith(dir))
+                    .filter(
+                        (file) =>
+                            file.workingDir != "U" && file.path.startsWith(dir)
+                    )
                     .map((file) => file.path);
             } else {
-                files = status.changed.map((file) => file.path);
+                files = status.changed
+                    .filter((file) => file.workingDir != "U")
+                    .map((file) => file.path);
             }
         } else {
-            files = (await this.getUnstagedFiles(dir)).map(({ path }) => path);
+            files = (await this.getUnstagedFiles(dir))
+                .filter((file) => file.type != "A")
+                .map(({ path }) => path);
         }
 
         try {
@@ -391,6 +412,34 @@ export class IsomorphicGit extends GitManager {
             this.plugin.displayError(error);
             throw error;
         }
+    }
+
+    async getUntrackedPaths(opts: {
+        path?: string;
+        status?: Status;
+    }): Promise<string[]> {
+        const untrackedPaths: string[] = [];
+        if (opts.status) {
+            for (const file of opts.status.changed) {
+                if (
+                    file.index == "U" &&
+                    file.workingDir === "U" &&
+                    file.path.startsWith(
+                        opts.path != undefined ? `${opts.path}/` : ""
+                    )
+                ) {
+                    untrackedPaths.push(file.path);
+                }
+            }
+        } else {
+            const status = await this.status({ path: opts?.path });
+            for (const file of status.changed) {
+                if (file.index === "U" && file.workingDir === "U") {
+                    untrackedPaths.push(file.path);
+                }
+            }
+        }
+        return untrackedPaths;
     }
 
     getProgressText(action: string, event: GitProgressEvent): string {
@@ -522,10 +571,12 @@ export class IsomorphicGit extends GitManager {
             ).length;
 
             this.plugin.setPluginState({ gitAction: CurrentGitAction.push });
+            const remote = await this.getCurrentRemote();
 
             await this.wrapFS(
                 git.push({
                     ...this.getRepo(),
+                    remote,
                     onProgress: (progress) => {
                         if (progressNotice !== undefined) {
                             progressNotice.noticeEl.innerText =
@@ -1039,14 +1090,20 @@ export class IsomorphicGit extends GitManager {
                         if (!workdirOid) {
                             return {
                                 path: filepath,
-                                deleted: true,
+                                type: "D",
+                            };
+                        }
+                        if (!stageOid) {
+                            return {
+                                path: filepath,
+                                type: "A",
                             };
                         }
 
                         if (workdirOid !== stageOid) {
                             return {
                                 path: filepath,
-                                deleted: false,
+                                type: "M",
                             };
                         }
                         return null;
@@ -1236,42 +1293,24 @@ function fromValue(value: any) {
     };
 }
 
-function getIterator(iterable: any) {
-    if (iterable[Symbol.asyncIterator]) {
-        return iterable[Symbol.asyncIterator]();
-    }
-    if (iterable[Symbol.iterator]) {
-        return iterable[Symbol.iterator]();
-    }
-    if (iterable.next) {
-        return iterable;
-    }
-    return fromValue(iterable);
+async function* arrayBufferToAsyncIterator(
+    buffer: ArrayBuffer
+): AsyncIterableIterator<Uint8Array> {
+    yield new Uint8Array(buffer);
 }
 
-async function forAwait(iterable: any, cb: any) {
-    const iter = getIterator(iterable);
-    while (true) {
-        const { value, done } = await iter.next();
-        if (value) await cb(value);
-        if (done) break;
-    }
-    if (iter.return) iter.return();
-}
-
-async function collect(iterable: any): Promise<Uint8Array> {
-    let size = 0;
-    const buffers: Uint8Array[] = [];
-    // This will be easier once `for await ... of` loops are available.
-    await forAwait(iterable, (value: any) => {
-        buffers.push(value);
-        size += value.byteLength;
+async function asyncIteratorToArrayBuffer(
+    iterator: AsyncIterableIterator<Uint8Array>
+): Promise<ArrayBuffer> {
+    const stream = new ReadableStream({
+        async start(controller) {
+            for await (const chunk of iterator) {
+                controller.enqueue(chunk);
+            }
+            controller.close();
+        },
     });
-    const result = new Uint8Array(size);
-    let nextIndex = 0;
-    for (const buffer of buffers) {
-        result.set(buffer, nextIndex);
-        nextIndex += buffer.byteLength;
-    }
-    return result;
+
+    const response = new Response(stream);
+    return await response.arrayBuffer();
 }