+++ src/setting/settings.ts
@@ -224,6 +224,20 @@ export class ObsidianGitSettingsTab extends PluginSettingTab {
                 });
 
             new Setting(containerEl)
+                .setName(`Auto ${commitOrSync} only staged files`)
+                .setDesc(
+                    `If turned on, only staged files are committed on ${commitOrSync}. If turned off, all changed files are committed.`
+                )
+                .addToggle((toggle) =>
+                    toggle
+                        .setValue(plugin.settings.autoCommitOnlyStaged)
+                        .onChange(async (value) => {
+                            plugin.settings.autoCommitOnlyStaged = value;
+                            await plugin.saveSettings();
+                        })
+                );
+
+            new Setting(containerEl)
                 .setName(
                     `Specify custom commit message on auto ${commitOrSync}`
                 )
@@ -591,7 +605,7 @@ export class ObsidianGitSettingsTab extends PluginSettingTab {
         new Setting(containerEl)
             .setName("Disable error notifications")
             .setDesc(
-                "Disable errror notifications of any kind to minimize distraction (refer to status bar for updates)."
+                "Disable error notifications of any kind to minimize distraction (refer to status bar for updates)."
             )
             .addToggle((toggle) =>
                 toggle
@@ -738,7 +752,7 @@ export class ObsidianGitSettingsTab extends PluginSettingTab {
         new Setting(containerEl)
             .setName("Advanced")
             .setDesc(
-                "These settings usually don't need to be changed, but may be requried for special setups."
+                "These settings usually don't need to be changed, but may be required for special setups."
             )
             .setHeading();
 
@@ -961,7 +975,7 @@ export class ObsidianGitSettingsTab extends PluginSettingTab {
     }
 
     /**
-     * Ensure, that certain last shown values are persisten in the settings.
+     * Ensure, that certain last shown values are persistent in the settings.
      *
      * Necessary for the line author info gutter context menus.
      */
@@ -1181,7 +1195,7 @@ export class ObsidianGitSettingsTab extends PluginSettingTab {
                 }).descEl.innerHTML = `
                     The CSS color of the gutter text.<br/>
 
-                    It is higly recommended to use
+                    It is highly recommended to use
                     <a href="https://developer.mozilla.org/en-US/docs/Web/CSS/Using_CSS_custom_properties">
                     CSS variables</a>
                     defined by themes
@@ -1322,7 +1336,7 @@ export class ObsidianGitSettingsTab extends PluginSettingTab {
         const defaultValue = DEFAULT_SETTINGS[settingsProperty];
 
         if (defaultValue !== storedValue) {
-            text.setValue(String(storedValue));
+            text.setValue(JSON.stringify(storedValue));
         }
     }