+++ src/types.ts
@@ -12,6 +12,7 @@ export interface ObsidianGitSettings {
     autoPushInterval: number;
     autoPullInterval: number;
     autoPullOnBoot: boolean;
+    autoCommitOnlyStaged: boolean;
     syncMethod: SyncMethod;
     resolutionMethod: ResolutionMethod;
     /**
@@ -263,10 +264,7 @@ export interface WalkDifference {
     type: "M" | "A" | "D";
 }
 
-export interface UnstagedFile {
-    path: string;
-    deleted: boolean;
-}
+export type UnstagedFile = WalkDifference;
 
 export interface BranchInfo {
     current?: string;
@@ -288,7 +286,7 @@ export type StatusRootTreeItem = RootTreeItem<FileStatusResult>;
 
 export type HistoryRootTreeItem = RootTreeItem<DiffFile>;
 
-export interface DiffViewState {
+export type DiffViewState = {
     /**
      * The repo relative file path for a.
      * For diffing a renamed file, this is the old path.
@@ -309,10 +307,10 @@ export interface DiffViewState {
     /**
      * The git ref to specify which state of that file should be shown.
      * An empty string refers to the index version of a file, so you have to specifically check against undefined.
-     * `undefined` stands for the workign tree version.
+     * `undefined` stands for the working tree version.
      */
     bRef?: string;
-}
+};
 
 export enum FileType {
     staged,