+++ src/ui/modals/discardModal.ts
@@ -1,45 +1,103 @@
 import type { App } from "obsidian";
 import { Modal } from "obsidian";
+import { plural } from "src/utils";
+
+export type DiscardResult = false | "delete" | "discard";
 
 export class DiscardModal extends Modal {
-    constructor(
-        app: App,
-        private readonly deletion: boolean,
-        private readonly filename: string
-    ) {
+    path: string;
+    deleteCount: number;
+    discardCount: number;
+    constructor({
+        app,
+        path,
+        filesToDeleteCount,
+        filesToDiscardCount,
+    }: {
+        app: App;
+        path: string;
+        filesToDeleteCount: number;
+        filesToDiscardCount: number;
+    }) {
         super(app);
+        this.path = path;
+        this.deleteCount = filesToDeleteCount;
+        this.discardCount = filesToDiscardCount;
     }
-    resolve: ((value: boolean | PromiseLike<boolean>) => void) | null = null;
-    myOpen() {
+    resolve: ((value: DiscardResult) => void) | null = null;
+
+    /**
+     * @returns the result of the modal, whcih can be:
+     *   - `false` if the user canceled the modal
+     *   - `"delete"` if the user chose to delete all files. In case there are also tracked files, they will be discarded as well.
+     *   - `"discard"` if the user chose to discard all tracked files. Untracked files will not be deleted.
+     */
+    openAndGetResult(): Promise<DiscardResult> {
         this.open();
-        return new Promise((resolve) => {
+        return new Promise<DiscardResult>((resolve) => {
             this.resolve = resolve;
         });
     }
+
     onOpen() {
+        const sum = this.deleteCount + this.discardCount;
         const { contentEl, titleEl } = this;
-        titleEl.setText(`${this.deletion ? "Delete" : "Discard"} this file?`);
-        contentEl
-            .createEl("p")
-            .setText(
-                `Do you really want to ${
-                    this.deletion ? "delete" : "discard the changes of"
-                } "${this.filename}"`
-            );
+        let titlePart = "";
+        if (this.path != "") {
+            if (sum > 1) {
+                titlePart = `files in "${this.path}"`;
+            } else {
+                titlePart = `"${this.path}"`;
+            }
+        }
+        titleEl.setText(
+            `${this.discardCount == 0 ? "Delete" : "Discard"} ${titlePart}`
+        );
+        if (this.deleteCount > 0) {
+            contentEl
+                .createEl("p")
+                .setText(
+                    `Are you sure you want to DELETE the ${plural(this.deleteCount, "untracked file")}? They are deleted according to your Obsidian trash settting.`
+                );
+        }
+        if (this.discardCount > 0) {
+            contentEl
+                .createEl("p")
+                .setText(
+                    `Are you sure you want to discard ALL changes in ${plural(this.discardCount, "tracked file")}?`
+                );
+        }
         const div = contentEl.createDiv({ cls: "modal-button-container" });
 
-        const discard = div.createEl("button", {
-            cls: "mod-warning",
-            text: this.deletion ? "Delete" : "Discard",
-        });
-        discard.addEventListener("click", () => {
-            if (this.resolve) this.resolve(true);
-            this.close();
-        });
-        discard.addEventListener("keypress", () => {
-            if (this.resolve) this.resolve(true);
-            this.close();
-        });
+        if (this.deleteCount > 0) {
+            const discardAndDelete = div.createEl("button", {
+                cls: "mod-warning",
+                text: `${this.discardCount > 0 ? "Discard" : "Delete"} all ${plural(sum, "file")}`,
+            });
+            discardAndDelete.addEventListener("click", () => {
+                if (this.resolve) this.resolve("delete");
+                this.close();
+            });
+            discardAndDelete.addEventListener("keypress", () => {
+                if (this.resolve) this.resolve("delete");
+                this.close();
+            });
+        }
+
+        if (this.discardCount > 0) {
+            const discard = div.createEl("button", {
+                cls: "mod-warning",
+                text: `Discard all ${plural(this.discardCount, "tracked file")}`,
+            });
+            discard.addEventListener("click", () => {
+                if (this.resolve) this.resolve("discard");
+                this.close();
+            });
+            discard.addEventListener("keypress", () => {
+                if (this.resolve) this.resolve("discard");
+                this.close();
+            });
+        }
 
         const close = div.createEl("button", {
             text: "Cancel",