+++ src/ui/sourceControl/components/stagedFileComponent.svelte
@@ -58,7 +58,8 @@
     function showDiff(event: MouseEvent) {
         event.stopPropagation();
         view.plugin.tools.openDiff({
-            aFile: change.path,
+            aFile: change.from ?? change.path,
+            bFile: change.path,
             aRef: "HEAD",
             bRef: "",
             event,
@@ -130,11 +131,3 @@
         </div>
     </div>
 </main>
-
-<style lang="scss">
-    main {
-        .nav-file-title {
-            align-items: center;
-        }
-    }
-</style>